backend/requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.1.2  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.5.0  # https://github.com/andymccurdy/redis-py
hiredis==1.0.1  # https://github.com/redis/hiredis-py
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
flower==0.9.4  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.9.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-compressor==2.4  # https://github.com/django-compressor/django-compressor
django-redis==4.11.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework

# changa
# ------------------------------------------------------------------------------
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
djangorestframework-jsonapi==3.1.0   # https://github.com/django-json-api/django-rest-framework-json-api
django-rest-framework-social-oauth2==1.1.0  # https://github.com/RealmTeam/django-rest-framework-social-oauth2
facebook-sdk==3.1.0  # https://github.com/mobolic/facebook-sdk
django-cors-headers==3.3.0  # https://github.com/adamchainz/django-cors-headers
qrcode==6.1
pymaging @ git+git://github.com/ojii/pymaging.git@596a08fce5664e58d6e8c96847393fbe987783f2
pymaging-png @ git+git://github.com/ojii/pymaging-png.git@83d85c44e4b2342818e6c068065e031a9f81bb9f
django-anymail==8.0
